feat: Add debug logging for Gnd and Velodrome exchangers
This commit adds debug logging for the Gnd and Velodrome exchangers. When the headless browser is started, a debug log message is printed, including the name of the exchanger. Similarly, when the browser is closed, another debug log message is printed. This logging will help with troubleshooting and monitoring the exchanger services.

The logging is added to the `getPoolsData` method in the `GndService` and `VelodromeService` classes.