feat(adapters): add CurveService to adapters
Add CurveService to the list of services used by AdaptersModule and AdaptersService, allowing for interaction with the Curve exchanger. This change modifies the `adapters.module.ts` and `adapters.service.ts` files.

The `adapters.module.ts` file was modified to import and include the new CurveService in the `providers` array.

The `adapters.service.ts` file was modified to include the CurveService in the class constructor, and to add a new condition for handling pools of type ExchangerType.CURVE. The `getPoolsData` method was also updated to call the new CurveService method.

A new file `curve.service.ts` was added to the `adapters/exchangers` folder. The `curve.service.ts` file contains the implementation of the CurveService class, which includes methods for loading pool data from the Curve API.

The `exchanger.service.ts` file was also modified, updating the logic in the `for` loop to only process exchanger_type of ExchangerType.CURVE.

The `exchanger.type.ts` file was updated to include a new enum value for ExchangerType.CURVE.

The `skim.service.ts` file was modified to comment out an error-throwing line without making any other changes.

---

This commit adds the CurveService to the Adapters module, allowing for interaction with the Curve exchanger. It also includes modifications to the AdaptersService and ExchangerService to handle the new service, as well as updates to some related files.